requirements.txt
numpy 
pandas 
tensorflow
tf-keras
matplotlib
seaborn
nltk
regex
scikit-learn
uvicorn==0.18.3
requests
transformers
pyYAML
types-pyYAML
python-box==6.0.2
ensure==1.0.2
joblib
tqdm
dill
-e .
